Replace CreateDoc file with my changes
